+++ database/migrations/2025_09_11_181212_create_companies_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('companies')) {
+            return;
+        }
+
         Schema::create('companies', function (Blueprint $table) {
             $table->id();
             $table->string('name');
@@ -26,6 +30,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('companies');
+        if (Schema::hasTable('companies')) {
+            Schema::dropIfExists('companies');
+        }
     }
 };
